refactor(local): Update Session interface and introduce AuthorizedSession type
Introduced `authorizedSession` type to extend the existing `session` type and added a new `Session` interface with methods for authentication and validation.
